+++ dataget/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.4.5"
+__version__ = "0.4.6"
 
 # datasets
 from .dataset import Dataset