+++ dataget/dataset.py
@@ -65,4 +65,3 @@ class Dataset(ABC):
 class DownloadError(Exception):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-