+++ dataget/vision/fashion_mnist.py
@@ -56,4 +56,3 @@ class fashion_mnist(Dataset):
             y_test = idx2numpy.convert_from_file(f)
 
         return X_train, y_train, X_test, y_test
-